Revert "ViewMap und AddGeoTag zusammengefasst"
This reverts commit 425208812184b465deca0e6bda8d33daa3dd37a4.

+++ test/GeoTagging/src/com/testrunns/geotagging/GeoTag.java
@@ -59,7 +59,7 @@ public class GeoTag {
 		return longitude;
 	}
 
-	public void setLongitude(double longitude) {
+	public void setLongitude(float longitude) {
 		this.longitude = longitude;
 	}
 
@@ -67,7 +67,7 @@ public class GeoTag {
 		return latitude;
 	}
 
-	public void setLatitude(double latitude) {
+	public void setLatitude(float latitude) {
 		this.latitude = latitude;
 	}
